Add tracking of version's server
When a version is downloaded, a server.ini file is created with the server that version was downloaded from. When downloading any additional components afterwards, it will reference this to switch back to that server.

This file can also be included prepacked to ensure installing the version as a zip still works.